ADD: Code refactor for better organization. Not productive yet
